Enemu controllers and Spell class update
So there is now a button used to create spell enums these enums will be used as an id so that objects can refernce the object, this was just an int but that confused me. the enum name will be the same as the class name.

added a few more animator controllers to the enemies and there are a few enemies placed in the test room